refactor: split workflow service into focused modules
