+++ backend/src/app/application/services.py
@@ -1,650 +1,4 @@
-from __future__ import annotations
+from app.application.exceptions import NotFoundError, ProviderNotConfiguredError
+from app.application.workflow_service import WorkflowService
 
-from datetime import UTC, datetime
-from typing import Callable
-from uuid import uuid4
-
-from pydantic import ValidationError
-from sqlalchemy import select
-from sqlalchemy.ext.asyncio import AsyncSession
-from sqlalchemy.orm import selectinload
-
-from app.api.schemas import (
-    AIActionResponse,
-    AISuggestionResponse,
-    CampaignCreate,
-    CampaignDetailResponse,
-    CampaignSummary,
-    ContentPieceCreate,
-    ContentPieceResponse,
-    ContentPieceUpdate,
-    DraftDecisionStatus,
-    DraftHistoryItemResponse,
-    GenerateDraftRequest,
-    MetadataPayload,
-    ProviderSettingsResponse,
-    ReviewActionResponse,
-    ReviewRequest,
-    ReviewResponse,
-    TranslationVersionResponse,
-    TranslateRequest,
-    UpdateProviderSettingsRequest,
-    WorkflowCounts,
-)
-from app.domain.enums import AISuggestionStatus, OperationType, ReviewActionType, ReviewState
-from app.config import Settings
-from app.infrastructure.ai.base import AIProvider, GeneratedPayload
-from app.infrastructure.ai.factory import build_ai_provider_for_selection
-from app.infrastructure.ai.metadata_parser import parse_metadata_output
-from app.infrastructure.ai.provider_settings_crypto import (
-    decrypt_api_key,
-    encrypt_api_key,
-    fingerprint_api_key,
-)
-from app.infrastructure.ai.translation_parser import parse_translation_output
-from app.infrastructure.db.models import AIProviderSettings, AISuggestion, Campaign, ContentPiece, ReviewAction
-from app.infrastructure.events.bus import EventBus
-
-
-class NotFoundError(ValueError):
-    pass
-
-
-class ProviderNotConfiguredError(ValueError):
-    pass
-
-
-GLOBAL_PROVIDER_SETTINGS_ID = "global"
-
-
-class WorkflowService:
-    def __init__(
-        self,
-        event_bus: EventBus,
-        ai_provider: AIProvider | None = None,
-        settings: Settings | None = None,
-        provider_builder: Callable[[str, str], AIProvider] | None = None,
-    ) -> None:
-        self.ai_provider = ai_provider
-        self.event_bus = event_bus
-        self.settings = settings
-        self.provider_builder = provider_builder
-
-    async def create_campaign(self, session: AsyncSession, payload: CampaignCreate) -> CampaignSummary:
-        now = datetime.now(UTC)
-        campaign = Campaign(
-            id=str(uuid4()),
-            name=payload.name,
-            description=payload.description,
-            created_at=now,
-            updated_at=now,
-        )
-        session.add(campaign)
-        await session.commit()
-        await session.refresh(campaign)
-        return CampaignSummary.model_validate(
-            {**campaign.__dict__, "content_piece_count": 0, "workflow_counts": self._workflow_counts([])}
-        )
-
-    async def list_campaigns(self, session: AsyncSession) -> list[CampaignSummary]:
-        result = await session.execute(
-            select(Campaign).options(selectinload(Campaign.content_pieces)).order_by(Campaign.created_at.desc())
-        )
-        campaigns = result.scalars().all()
-        return [
-            CampaignSummary.model_validate(
-                {
-                    **campaign.__dict__,
-                    "content_piece_count": len(campaign.content_pieces),
-                    "workflow_counts": self._workflow_counts(campaign.content_pieces),
-                }
-            )
-            for campaign in campaigns
-        ]
-
-    async def get_provider_settings(self, session: AsyncSession) -> ProviderSettingsResponse:
-        stored_settings = await session.get(AIProviderSettings, GLOBAL_PROVIDER_SETTINGS_ID)
-        if stored_settings is not None:
-            return ProviderSettingsResponse(
-                provider=stored_settings.provider,
-                configured=True,
-                has_api_key=True,
-                source="database",
-            )
-
-        provider, api_key = self._get_environment_provider_config()
-        if provider and api_key:
-            return ProviderSettingsResponse(
-                provider=provider,
-                configured=True,
-                has_api_key=True,
-                source="environment",
-            )
-
-        return ProviderSettingsResponse(
-            provider=None,
-            configured=False,
-            has_api_key=False,
-            source="missing",
-        )
-
-    async def update_provider_settings(
-        self,
-        session: AsyncSession,
-        payload: UpdateProviderSettingsRequest,
-    ) -> ProviderSettingsResponse:
-        if self.settings is None:
-            raise ValueError("Settings-backed provider configuration is unavailable.")
-
-        stored_settings = await session.get(AIProviderSettings, GLOBAL_PROVIDER_SETTINGS_ID)
-        api_key = (payload.api_key or "").strip()
-        provider_changed = stored_settings is not None and stored_settings.provider != payload.provider
-
-        if stored_settings is None and not api_key:
-            raise ValueError("api_key is required when no database provider settings exist.")
-        if provider_changed and not api_key:
-            raise ValueError("api_key is required when switching provider.")
-
-        now = datetime.now(UTC)
-        if stored_settings is None:
-            stored_settings = AIProviderSettings(
-                id=GLOBAL_PROVIDER_SETTINGS_ID,
-                provider=payload.provider,
-                encrypted_api_key=encrypt_api_key(api_key, self.settings.ai_settings_encryption_key),
-                api_key_fingerprint=fingerprint_api_key(api_key),
-                created_at=now,
-                updated_at=now,
-            )
-            session.add(stored_settings)
-        else:
-            stored_settings.provider = payload.provider
-            if api_key:
-                stored_settings.encrypted_api_key = encrypt_api_key(api_key, self.settings.ai_settings_encryption_key)
-                stored_settings.api_key_fingerprint = fingerprint_api_key(api_key)
-            stored_settings.updated_at = now
-
-        await session.commit()
-        return ProviderSettingsResponse(
-            provider=stored_settings.provider,
-            configured=True,
-            has_api_key=True,
-            source="database",
-        )
-
-    async def get_campaign(self, session: AsyncSession, campaign_id: str) -> CampaignDetailResponse:
-        result = await session.execute(
-            select(Campaign)
-            .where(Campaign.id == campaign_id)
-            .options(
-                selectinload(Campaign.content_pieces).selectinload(ContentPiece.ai_suggestions),
-                selectinload(Campaign.content_pieces).selectinload(ContentPiece.review_actions),
-            )
-        )
-        campaign = result.scalar_one_or_none()
-        if campaign is None:
-            raise NotFoundError("Campaign not found")
-        content_pieces = [self._serialize_content_piece(piece) for piece in campaign.content_pieces]
-        return CampaignDetailResponse(
-            id=campaign.id,
-            name=campaign.name,
-            description=campaign.description,
-            created_at=campaign.created_at,
-            updated_at=campaign.updated_at,
-            workflow_counts=self._workflow_counts(campaign.content_pieces),
-            content_pieces=content_pieces,
-        )
-
-    async def create_content_piece(
-        self, session: AsyncSession, campaign_id: str, payload: ContentPieceCreate
-    ) -> ContentPieceResponse:
-        campaign = await session.get(Campaign, campaign_id)
-        if campaign is None:
-            raise NotFoundError("Campaign not found")
-
-        now = datetime.now(UTC)
-        piece = ContentPiece(
-            id=str(uuid4()),
-            campaign_id=campaign_id,
-            type="content",
-            source_text=payload.source_text,
-            current_text=payload.current_text or payload.source_text,
-            source_language=None,
-            target_language=None,
-            review_state=ReviewState.DRAFT.value,
-            created_at=now,
-            updated_at=now,
-        )
-        session.add(piece)
-        await session.commit()
-        piece = await self._load_content_piece(session, piece.id)
-        await self._publish_content_piece_event(piece, "content_piece.updated")
-        return self._serialize_content_piece(piece)
-
-    async def get_content_piece(self, session: AsyncSession, content_piece_id: str) -> ContentPieceResponse:
-        piece = await self._load_content_piece(session, content_piece_id)
-        return self._serialize_content_piece(piece)
-
-    async def update_content_piece(
-        self, session: AsyncSession, content_piece_id: str, payload: ContentPieceUpdate
-    ) -> ContentPieceResponse:
-        piece = await self._load_content_piece(session, content_piece_id)
-
-        changed = False
-        for field in ("type", "source_text", "current_text", "source_language", "target_language", "review_state"):
-            value = getattr(payload, field)
-            if value is not None and getattr(piece, field) != value:
-                setattr(piece, field, value.value if isinstance(value, ReviewState) else value)
-                changed = True
-
-        if changed:
-            piece.updated_at = datetime.now(UTC)
-            await session.commit()
-            piece = await self._load_content_piece(session, piece.id)
-            await self._publish_content_piece_event(piece, "content_piece.updated")
-        return self._serialize_content_piece(piece)
-
-    async def generate_draft(
-        self, session: AsyncSession, content_piece_id: str, payload: GenerateDraftRequest
-    ) -> AIActionResponse:
-        piece = await self._load_content_piece(session, content_piece_id)
-        ai_provider = await self._resolve_ai_provider(session)
-        generated = await self._safe_ai_call(
-            ai_provider=ai_provider,
-            operation=OperationType.GENERATE_DRAFT,
-            invoke=lambda: ai_provider.generate_draft(
-                source_text=piece.current_text,
-                content_type=piece.type,
-                context=payload.context,
-            ),
-        )
-        suggestion, piece = await self._persist_suggestion(
-            session, piece, generated, OperationType.GENERATE_DRAFT, ai_provider=ai_provider
-        )
-        return AIActionResponse(
-            suggestion=self._serialize_suggestion(suggestion),
-            content_piece=self._serialize_content_piece(piece),
-        )
-
-    async def translate(
-        self, session: AsyncSession, content_piece_id: str, payload: TranslateRequest
-    ) -> AIActionResponse:
-        piece = await self._load_content_piece(session, content_piece_id)
-        ai_provider = await self._resolve_ai_provider(session)
-        generated = await self._safe_ai_call(
-            ai_provider=ai_provider,
-            operation=OperationType.TRANSLATE,
-            invoke=lambda: ai_provider.translate(
-                source_text=piece.current_text,
-                source_language=payload.source_language,
-                target_language=payload.target_language,
-                context=payload.context,
-            ),
-        )
-        piece.source_language = payload.source_language
-        piece.target_language = payload.target_language
-        suggestion, piece = await self._persist_suggestion(
-            session,
-            piece,
-            generated,
-            OperationType.TRANSLATE,
-            ai_provider=ai_provider,
-            source_language=payload.source_language,
-            target_language=payload.target_language,
-        )
-        return AIActionResponse(
-            suggestion=self._serialize_suggestion(suggestion),
-            content_piece=self._serialize_content_piece(piece),
-        )
-
-    async def extract_metadata(self, session: AsyncSession, content_piece_id: str) -> AIActionResponse:
-        piece = await self._load_content_piece(session, content_piece_id)
-        ai_provider = await self._resolve_ai_provider(session)
-        generated = await self._safe_ai_call(
-            ai_provider=ai_provider,
-            operation=OperationType.EXTRACT_METADATA,
-            invoke=lambda: ai_provider.extract_metadata(
-                source_text=piece.current_text,
-                content_type=piece.type,
-            ),
-        )
-        suggestion, piece = await self._persist_suggestion(
-            session, piece, generated, OperationType.EXTRACT_METADATA, ai_provider=ai_provider
-        )
-        return AIActionResponse(
-            suggestion=self._serialize_suggestion(suggestion),
-            content_piece=self._serialize_content_piece(piece),
-        )
-
-    async def review(
-        self, session: AsyncSession, content_piece_id: str, payload: ReviewRequest
-    ) -> ReviewResponse:
-        piece = await self._load_content_piece(session, content_piece_id)
-        suggestion = None
-        if payload.ai_suggestion_id:
-            suggestion = await session.get(AISuggestion, payload.ai_suggestion_id)
-            if suggestion is None or suggestion.content_piece_id != piece.id:
-                raise NotFoundError("Suggestion not found for content piece")
-
-        if payload.action == ReviewActionType.EDIT and not payload.edited_text:
-            raise ValueError("edited_text is required for edit actions")
-        if payload.action in {ReviewActionType.ACCEPT, ReviewActionType.REJECT} and suggestion is None:
-            raise ValueError("ai_suggestion_id is required for accept and reject actions")
-        if (
-            suggestion is not None
-            and payload.action in {ReviewActionType.ACCEPT, ReviewActionType.REJECT}
-            and suggestion.operation_type != OperationType.GENERATE_DRAFT.value
-        ):
-            raise ValueError("Only generated drafts can be accepted or rejected")
-        if (
-            suggestion is not None
-            and payload.action in {ReviewActionType.ACCEPT, ReviewActionType.REJECT}
-            and suggestion.status != AISuggestionStatus.SUCCESS.value
-        ):
-            raise ValueError("Only successful draft suggestions can be accepted or rejected")
-
-        if payload.action == ReviewActionType.ACCEPT:
-            piece.current_text = (suggestion.output_text if suggestion else piece.current_text) or piece.current_text
-        elif payload.action == ReviewActionType.EDIT:
-            piece.current_text = payload.edited_text or piece.current_text
-
-        piece.updated_at = datetime.now(UTC)
-
-        review_action = ReviewAction(
-            id=str(uuid4()),
-            content_piece_id=piece.id,
-            ai_suggestion_id=payload.ai_suggestion_id,
-            action=payload.action.value,
-            comment=payload.comment,
-            edited_text=payload.edited_text,
-            created_at=datetime.now(UTC),
-        )
-        session.add(review_action)
-        await session.commit()
-        await session.refresh(review_action)
-        piece = await self._load_content_piece(session, piece.id)
-
-        await self._publish_review_event(piece, review_action)
-        return ReviewResponse(
-            review_action=ReviewActionResponse.model_validate(review_action),
-            content_piece=self._serialize_content_piece(piece),
-        )
-
-    async def _safe_ai_call(
-        self,
-        *,
-        ai_provider: AIProvider,
-        operation: OperationType,
-        invoke,
-    ) -> tuple[GeneratedPayload | None, AISuggestionStatus, str | None, dict | None]:
-        try:
-            generated: GeneratedPayload = await invoke()
-            structured_output = generated.structured_output
-            if operation == OperationType.EXTRACT_METADATA:
-                if structured_output is None and generated.output_text:
-                    structured_output = parse_metadata_output(generated.output_text)
-                if structured_output is not None:
-                    structured_output = MetadataPayload.model_validate(structured_output).model_dump()
-            if operation == OperationType.TRANSLATE and generated.output_text:
-                generated.output_text = parse_translation_output(generated.output_text)
-            return generated, AISuggestionStatus.SUCCESS, generated.output_text, structured_output
-        except (ValidationError, ValueError) as exc:
-            return None, AISuggestionStatus.FAILED, str(exc), None
-        except Exception as exc:  # pragma: no cover - external provider failure
-            return None, AISuggestionStatus.FAILED, str(exc), None
-
-    async def _persist_suggestion(
-        self,
-        session: AsyncSession,
-        piece: ContentPiece,
-        generated_result: tuple[GeneratedPayload | None, AISuggestionStatus, str | None, dict | None],
-        operation_type: OperationType,
-        ai_provider: AIProvider,
-        source_language: str | None = None,
-        target_language: str | None = None,
-    ) -> tuple[AISuggestion, ContentPiece]:
-        _, status, output_text, structured_output = generated_result
-        suggestion = AISuggestion(
-            id=str(uuid4()),
-            content_piece_id=piece.id,
-            provider=ai_provider.provider_name,
-            model=ai_provider.model_name,
-            operation_type=operation_type.value,
-            input_text=piece.current_text,
-            output_text=output_text,
-            source_language=source_language,
-            target_language=target_language,
-            structured_output_json=structured_output,
-            status=status.value,
-            created_at=datetime.now(UTC),
-        )
-        session.add(suggestion)
-        await session.commit()
-        piece = await self._load_content_piece(session, piece.id)
-        fresh_suggestion = await session.get(AISuggestion, suggestion.id)
-        await self._publish_ai_event(piece, fresh_suggestion)
-        return fresh_suggestion, piece
-
-    async def _load_content_piece(self, session: AsyncSession, content_piece_id: str) -> ContentPiece:
-        result = await session.execute(
-            select(ContentPiece)
-            .where(ContentPiece.id == content_piece_id)
-            .execution_options(populate_existing=True)
-            .options(
-                selectinload(ContentPiece.ai_suggestions),
-                selectinload(ContentPiece.review_actions),
-            )
-        )
-        piece = result.scalar_one_or_none()
-        if piece is None:
-            raise NotFoundError("Content piece not found")
-        return piece
-
-    def _serialize_content_piece(self, piece: ContentPiece) -> ContentPieceResponse:
-        latest_suggestion_model = None
-        latest_reviewable_suggestion_model = None
-        latest_review_model = None
-        latest_metadata_attempt_model = None
-        latest_metadata = None
-        draft_history: list[DraftHistoryItemResponse] = []
-        translation_versions: list[TranslationVersionResponse] = []
-        ai_call_history: list[AISuggestionResponse] = []
-
-        latest_suggestion = max(piece.ai_suggestions, key=lambda item: item.created_at, default=None)
-        latest_review = max(piece.review_actions, key=lambda item: item.created_at, default=None)
-        reviewable_suggestions = [
-            item
-            for item in piece.ai_suggestions
-            if item.operation_type == OperationType.GENERATE_DRAFT.value
-            and item.status == AISuggestionStatus.SUCCESS.value
-            and item.output_text
-        ]
-        latest_reviewable_suggestion = max(reviewable_suggestions, key=lambda item: item.created_at, default=None)
-        decision_by_suggestion_id = self._draft_decisions_by_suggestion(piece.review_actions)
-
-        if latest_suggestion is not None:
-            latest_suggestion_model = self._serialize_suggestion(latest_suggestion)
-        if latest_reviewable_suggestion is not None:
-            latest_reviewable_suggestion_model = self._serialize_suggestion(latest_reviewable_suggestion)
-        if latest_review is not None:
-            latest_review_model = ReviewActionResponse.model_validate(latest_review)
-        draft_history = [
-            DraftHistoryItemResponse(
-                id=item.id,
-                provider=item.provider,
-                model=item.model,
-                input_text=item.input_text,
-                output_text=self._normalize_suggestion_output(item) or "",
-                created_at=item.created_at,
-                decision_status=decision_by_suggestion_id.get(item.id, DraftDecisionStatus.PENDING),
-            )
-            for item in sorted(reviewable_suggestions, key=lambda item: item.created_at, reverse=True)
-        ]
-
-        metadata_suggestions = [
-            item
-            for item in piece.ai_suggestions
-            if item.operation_type == OperationType.EXTRACT_METADATA.value
-        ]
-        if metadata_suggestions:
-            latest_metadata_attempt = max(metadata_suggestions, key=lambda item: item.created_at)
-            latest_metadata_attempt_model = self._serialize_suggestion(latest_metadata_attempt)
-            if (
-                latest_metadata_attempt.status == AISuggestionStatus.SUCCESS.value
-                and latest_metadata_attempt.structured_output_json
-            ):
-                latest_metadata = MetadataPayload.model_validate(latest_metadata_attempt.structured_output_json)
-
-        translation_suggestions = sorted(
-            [
-                item
-                for item in piece.ai_suggestions
-                if item.operation_type == OperationType.TRANSLATE.value and item.output_text
-            ],
-            key=lambda item: item.created_at,
-            reverse=True,
-        )
-        translation_versions = [
-            TranslationVersionResponse.model_validate(
-                {
-                    **item.__dict__,
-                    "output_text": self._normalize_suggestion_output(item),
-                }
-            )
-            for item in translation_suggestions
-        ]
-        ai_call_history = [
-            self._serialize_suggestion(item)
-            for item in sorted(piece.ai_suggestions, key=lambda item: item.created_at)
-        ]
-
-        return ContentPieceResponse(
-            id=piece.id,
-            campaign_id=piece.campaign_id,
-            type=piece.type,
-            source_text=piece.source_text,
-            current_text=piece.current_text,
-            source_language=piece.source_language,
-            target_language=piece.target_language,
-            review_state=ReviewState(piece.review_state),
-            created_at=piece.created_at,
-            updated_at=piece.updated_at,
-            latest_suggestion=latest_suggestion_model,
-            latest_reviewable_suggestion=latest_reviewable_suggestion_model,
-            latest_review_action=latest_review_model,
-            latest_metadata_attempt=latest_metadata_attempt_model,
-            latest_metadata=latest_metadata,
-            draft_history=draft_history,
-            translation_versions=translation_versions,
-            ai_call_history=ai_call_history,
-        )
-
-    def _serialize_suggestion(self, suggestion: AISuggestion) -> AISuggestionResponse:
-        return AISuggestionResponse.model_validate(
-            {
-                **suggestion.__dict__,
-                "output_text": self._normalize_suggestion_output(suggestion),
-            }
-        )
-
-    def _normalize_suggestion_output(self, suggestion: AISuggestion) -> str | None:
-        if not suggestion.output_text:
-            return suggestion.output_text
-        if suggestion.operation_type == OperationType.TRANSLATE.value:
-            return parse_translation_output(suggestion.output_text)
-        return suggestion.output_text
-
-    def _draft_decisions_by_suggestion(self, review_actions: list[ReviewAction]) -> dict[str, DraftDecisionStatus]:
-        decisions: dict[str, DraftDecisionStatus] = {}
-        for action in sorted(review_actions, key=lambda item: item.created_at):
-            if not action.ai_suggestion_id:
-                continue
-            if action.action in {ReviewActionType.ACCEPT.value, ReviewActionType.EDIT.value}:
-                decisions[action.ai_suggestion_id] = DraftDecisionStatus.ACCEPTED
-            elif action.action == ReviewActionType.REJECT.value:
-                decisions[action.ai_suggestion_id] = DraftDecisionStatus.REJECTED
-        return decisions
-
-    async def _resolve_ai_provider(self, session: AsyncSession) -> AIProvider:
-        if self.ai_provider is not None:
-            return self.ai_provider
-        if self.settings is None:
-            raise ProviderNotConfiguredError("AI provider is not configured.")
-
-        stored_settings = await session.get(AIProviderSettings, GLOBAL_PROVIDER_SETTINGS_ID)
-        if stored_settings is not None:
-            api_key = decrypt_api_key(
-                stored_settings.encrypted_api_key,
-                self.settings.ai_settings_encryption_key,
-            )
-            return self._build_provider(stored_settings.provider, api_key)
-
-        provider, api_key = self._get_environment_provider_config()
-        if provider and api_key:
-            return self._build_provider(provider, api_key)
-
-        raise ProviderNotConfiguredError("AI provider is not configured.")
-
-    def _get_environment_provider_config(self) -> tuple[str | None, str | None]:
-        if self.settings is None:
-            return None, None
-        if self.settings.ai_provider == "openai" and self.settings.openai_api_key:
-            return "openai", self.settings.openai_api_key
-        if self.settings.ai_provider == "gemini" and self.settings.gemini_api_key:
-            return "gemini", self.settings.gemini_api_key
-        return None, None
-
-    def _build_provider(self, provider_name: str, api_key: str) -> AIProvider:
-        if self.provider_builder is not None:
-            return self.provider_builder(provider_name, api_key)
-        if self.settings is None:
-            raise ProviderNotConfiguredError("AI provider is not configured.")
-        return build_ai_provider_for_selection(self.settings, provider_name, api_key)
-
-    async def _publish_content_piece_event(self, piece: ContentPiece, event_type: str) -> None:
-        await self.event_bus.publish(
-            {
-                "type": event_type,
-                "campaign_id": piece.campaign_id,
-                "content_piece_id": piece.id,
-                "timestamp": datetime.now(UTC).isoformat(),
-                "payload": {"review_state": piece.review_state},
-            }
-        )
-
-    async def _publish_ai_event(self, piece: ContentPiece, suggestion: AISuggestion) -> None:
-        await self.event_bus.publish(
-            {
-                "type": "ai_suggestion.created",
-                "campaign_id": piece.campaign_id,
-                "content_piece_id": piece.id,
-                "timestamp": datetime.now(UTC).isoformat(),
-                "payload": {
-                    "suggestion_id": suggestion.id,
-                    "operation_type": suggestion.operation_type,
-                    "status": suggestion.status,
-                },
-            }
-        )
-        await self._publish_content_piece_event(piece, "content_piece.updated")
-
-    async def _publish_review_event(self, piece: ContentPiece, review_action: ReviewAction) -> None:
-        await self.event_bus.publish(
-            {
-                "type": "review_action.created",
-                "campaign_id": piece.campaign_id,
-                "content_piece_id": piece.id,
-                "timestamp": datetime.now(UTC).isoformat(),
-                "payload": {
-                    "review_action_id": review_action.id,
-                    "action": review_action.action,
-                    "review_state": piece.review_state,
-                },
-            }
-        )
-        await self._publish_content_piece_event(piece, "content_piece.updated")
-
-    def _workflow_counts(self, pieces: list[ContentPiece]) -> WorkflowCounts:
-        counts = WorkflowCounts()
-        for piece in pieces:
-            setattr(counts, piece.review_state, getattr(counts, piece.review_state) + 1)
-        return counts
+__all__ = ["NotFoundError", "ProviderNotConfiguredError", "WorkflowService"]
